# Accounting visited instructions with missing shapes