Blindly pushing to the worklist can lead to an infinite growth.
This change assures, that "undefined shape" instructions are only visited once (until they are repushed as a user of a newly shape-computed instruction in 'addDependentValuesToWL').

Assuming you'd want an example of this happening, I'll prepare a gist of the IR. (Which has only 800-ish lines of IR-code, but doesn't stop even after 3 million mWorklist size)