@simoll 2018-07-31
Thanks for reporting! I've committed a solution that is even faster. It encapsulates the worklist and enforces that an instruction can not be added to the worklist if it's already on it (d13a9c4).
Some stats on https://github.com/cdl-saarland/rv/blob/develop/test/suite/test_069_md5_pacxx-wfv.cpp with `time rvTool -wfv -lower -i build/test_069_md5_pacxx-wfv.ll -o  -k foo -s  -w 8 [..]`:
```

Before:   13.87s user 0.02s system 99% cpu 13.897 total
This PR:  0.15s  user 0.01s system 99% cpu 0.162 total
Commit:   0.08s  user 0.01s system 99% cpu 0.088 total
```